# wi-app/src/imports/openapi.json
{"openapi":"3.1.0","info":{"title":"SmartHome NLP API","description":"Pipeline NLP → RDF → GraphDB pentru descrieri Smart Home","version":"0.1.0"},"paths":{"/process":{"post":{"tags":["Process"],"summary":"Procesează fișier .txt","description":"Primește un fișier .txt, rulează pipeline-ul NLP complet și returnează\nentitățile, relațiile și numărul de triple RDF generate.","operationId":"process_file_process_post","parameters":[{"name":"upload","in":"query","required":false,"schema":{"type":"boolean","description":"Uploadează graful în GraphDB după procesare","default":true,"title":"Upload"},"description":"Uploadează graful în GraphDB după procesare"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_file_process_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/process/text":{"post":{"tags":["Process"],"summary":"Procesează text direct","description":"Primește un text în limbaj natural direct în body, rulează pipeline-ul\nNLP complet și returnează entitățile, relațiile și triple count.","operationId":"process_text_process_text_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/turtle":{"get":{"tags":["Graph"],"summary":"Returnează graful curent ca Turtle","description":"Returnează conținutul Turtle al grafului generat de ultimul run al pipeline-ului.","operationId":"get_turtle_graph_turtle_get","responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}},"text/turtle":{}}}}}},"/graph/upload":{"post":{"tags":["Graph"],"summary":"Uploadează graful în GraphDB","description":"Uploadează graful curent (din ultimul run al pipeline-ului) în GraphDB.","operationId":"upload_graph_graph_upload_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}}}}},"/sparql":{"post":{"tags":["SPARQL"],"summary":"Execută query SPARQL custom","description":"Execută un query SPARQL SELECT arbitrar pe GraphDB și returnează rezultatele\nîn format JSON (structura W3C: bindings list).","operationId":"run_custom_sparql_sparql_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SparqlRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SparqlResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sparql/predefined/{query_id}":{"get":{"tags":["SPARQL"],"summary":"Execută interogare predefinită","description":"Execută una din cele 6 interogări SPARQL predefinite.\n\n- **1** — Actuatoare în LivingRoom\n- **2** — Actuatoare cu brand și preț (date externe OntoRefine)\n- **3** — Pași din rutină și secvența lor (hasNext)\n- **4** — Acțiuni și actuatoarele folosite\n- **5** — Camere unde au loc acțiuni\n- **6** — Integrare completă: acțiune + actuator + cameră + preț","operationId":"run_predefined_sparql_predefined__query_id__get","parameters":[{"name":"query_id","in":"path","required":true,"schema":{"type":"integer","maximum":6,"minimum":1,"description":"ID-ul interogării predefinite (1–6)","title":"Query Id"},"description":"ID-ul interogării predefinite (1–6)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredefinedQueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entities":{"get":{"tags":["Extraction"],"summary":"Entitățile extrase","description":"Returnează toate entitățile extrase în ultimul run al pipeline-ului.\nDisponibile după cel puțin un apel la POST /process sau POST /process/text.","operationId":"get_entities_entities_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntitiesResponse"}}}}}}},"/relations":{"get":{"tags":["Extraction"],"summary":"Relațiile extrase","description":"Returnează toate relațiile extrase în ultimul run al pipeline-ului.\nDisponibile după cel puțin un apel la POST /process sau POST /process/text.","operationId":"get_relations_relations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelationsResponse"}}}}}}},"/health":{"get":{"tags":["Infra"],"summary":"Health Check","description":"Verifică că serviciul este activ.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"Body_process_file_process_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File","description":"Fișier text (.txt) cu descriere Smart Home"}},"type":"object","required":["file"],"title":"Body_process_file_process_post"},"EntitiesResponse":{"properties":{"source":{"type":"string","title":"Source"},"total":{"type":"integer","title":"Total"},"entities":{"items":{"$ref":"#/components/schemas/app__routers__entities__EntityOut"},"type":"array","title":"Entities"}},"type":"object","required":["source","total","entities"],"title":"EntitiesResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredefinedQueryResponse":{"properties":{"query_id":{"type":"integer","title":"Query Id"},"query_name":{"type":"string","title":"Query Name"},"description":{"type":"string","title":"Description"},"result_count":{"type":"integer","title":"Result Count"},"results":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Results"}},"type":"object","required":["query_id","query_name","description","result_count","results"],"title":"PredefinedQueryResponse"},"ProcessResponse":{"properties":{"source":{"type":"string","title":"Source"},"entity_count":{"type":"integer","title":"Entity Count"},"relation_count":{"type":"integer","title":"Relation Count"},"triple_count":{"type":"integer","title":"Triple Count"},"graphdb_uploaded":{"type":"boolean","title":"Graphdb Uploaded"},"entities":{"items":{"$ref":"#/components/schemas/app__routers__process__EntityOut"},"type":"array","title":"Entities"},"relations":{"items":{"$ref":"#/components/schemas/app__routers__process__RelationOut"},"type":"array","title":"Relations"},"ttl_path":{"type":"string","title":"Ttl Path"}},"type":"object","required":["source","entity_count","relation_count","triple_count","graphdb_uploaded","entities","relations","ttl_path"],"title":"ProcessResponse"},"RelationsResponse":{"properties":{"source":{"type":"string","title":"Source"},"total":{"type":"integer","title":"Total"},"relations":{"items":{"$ref":"#/components/schemas/app__routers__entities__RelationOut"},"type":"array","title":"Relations"}},"type":"object","required":["source","total","relations"],"title":"RelationsResponse"},"SparqlRequest":{"properties":{"query":{"type":"string","title":"Query"},"repository":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Repository"}},"type":"object","required":["query"],"title":"SparqlRequest"},"SparqlResponse":{"properties":{"repository":{"type":"string","title":"Repository"},"query":{"type":"string","title":"Query"},"result_count":{"type":"integer","title":"Result Count"},"results":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Results"}},"type":"object","required":["repository","query","result_count","results"],"title":"SparqlResponse"},"TextInput":{"properties":{"text":{"type":"string","title":"Text"},"source_name":{"type":"string","title":"Source Name","default":"inline"},"upload":{"type":"boolean","title":"Upload","default":true}},"type":"object","required":["text"],"title":"TextInput"},"UploadResponse":{"properties":{"repository":{"type":"string","title":"Repository"},"triple_count":{"type":"integer","title":"Triple Count"},"message":{"type":"string","title":"Message"}},"type":"object","required":["repository","triple_count","message"],"title":"UploadResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"app__routers__entities__EntityOut":{"properties":{"label":{"type":"string","title":"Label"},"uri_name":{"type":"string","title":"Uri Name"},"entity_type":{"type":"string","title":"Entity Type"},"confidence":{"type":"number","title":"Confidence"},"source_text":{"type":"string","title":"Source Text"}},"type":"object","required":["label","uri_name","entity_type","confidence","source_text"],"title":"EntityOut"},"app__routers__entities__RelationOut":{"properties":{"subject":{"type":"string","title":"Subject"},"predicate":{"type":"string","title":"Predicate"},"object":{"type":"string","title":"Object"},"confidence":{"type":"number","title":"Confidence"},"source_text":{"type":"string","title":"Source Text"}},"type":"object","required":["subject","predicate","object","confidence","source_text"],"title":"RelationOut"},"app__routers__process__EntityOut":{"properties":{"label":{"type":"string","title":"Label"},"uri_name":{"type":"string","title":"Uri Name"},"entity_type":{"type":"string","title":"Entity Type"},"confidence":{"type":"number","title":"Confidence"}},"type":"object","required":["label","uri_name","entity_type","confidence"],"title":"EntityOut"},"app__routers__process__RelationOut":{"properties":{"subject":{"type":"string","title":"Subject"},"predicate":{"type":"string","title":"Predicate"},"object":{"type":"string","title":"Object"},"confidence":{"type":"number","title":"Confidence"}},"type":"object","required":["subject","predicate","object","confidence"],"title":"RelationOut"}}}}